requirements.txt
Flask==2.1.2
gunicorn==20.1.0
requests
azure-storage-blob
